Add service deactivation functionality for Phone line. Add listing of available phone plans (draft)

+++ Others/Study_Lina/PersonalCabinet/src/main/java/com/perscab/model/Hardware.java
@@ -4,23 +4,23 @@ import java.math.BigInteger;
 
 public class Hardware {
 
-    private BigInteger serviceTypeId;
+    private long serviceTypeId;
     private String name;
     private String serialNumber;
     private String status;
 
-    public Hardware(BigInteger serviceTypeId, String name, String serialNumber, String status) {
+    public Hardware(long serviceTypeId, String name, String serialNumber, String status) {
         this.serviceTypeId = serviceTypeId;
         this.name = name;
         this.serialNumber = serialNumber;
         this.status = status;
     }
 
-    public BigInteger getServiceTypeId() {
+    public long getServiceTypeId() {
         return serviceTypeId;
     }
 
-    public void setServiceTypeId(BigInteger serviceTypeId) {
+    public void setServiceTypeId(long serviceTypeId) {
         this.serviceTypeId = serviceTypeId;
     }
